+++ python/tests/test_metadata.py
@@ -10,10 +10,11 @@ import numpy
 import pytest
 import inspect
 from pathlib import Path
+from os import path
 
 from ouster import client
 
-from tests.conftest import METADATA_DATA_DIR
+from tests.conftest import METADATA_DATA_DIR, PCAPS_DATA_DIR
 
 
 @pytest.mark.parametrize("mode, string", [
@@ -103,6 +104,13 @@ def test_read_info(meta: client.SensorInfo) -> None:
     assert meta.udp_port_lidar == 0
     assert meta.udp_port_imu == 0
 
+    assert meta.build_date == "2020-10-23T14:05:18Z"
+    assert meta.prod_pn == "840-102146-C"
+    assert meta.image_rev == "ousteros-image-prod-aries-v2.0.0-rc.2+20201023140416.staging"
+    assert meta.status == "RUNNING"
+    assert meta.cal == client.SensorCalibration()
+    assert meta.config == client.SensorConfig()
+
 
 def test_write_info(meta: client.SensorInfo) -> None:
     """Check modifying metadata."""
@@ -129,8 +137,16 @@ def test_write_info(meta: client.SensorInfo) -> None:
     meta.init_id = 0
     meta.udp_port_lidar = 0
     meta.udp_port_imu = 0
+    meta.build_date = ""
+    meta.image_rev = ""
+    meta.prod_pn = ""
+    meta.status = ""
+    meta.cal = client.SensorCalibration()
+    meta.config = client.SensorConfig()
 
-    assert meta == client.SensorInfo()
+    assert meta != client.SensorInfo()
+    assert meta.has_fields_equal(client.SensorInfo())
+    assert meta.original_string() != client.SensorInfo().original_string()
 
     with pytest.raises(TypeError):
         meta.mode = 1  # type: ignore
@@ -207,7 +223,7 @@ def test_info_length() -> None:
     info_attributes = inspect.getmembers(client.SensorInfo, lambda a: not inspect.isroutine(a))
     info_properties = [a for a in info_attributes if not (a[0].startswith('__') and a[0].endswith('__'))]
 
-    assert len(info_properties) == 16, "Don't forget to update tests and the sensor_info == operator!"
+    assert len(info_properties) == 22, "Don't forget to update tests and the sensor_info == operator!"
 
 
 def test_equality_format() -> None:
@@ -238,3 +254,30 @@ def test_skip_metadata_beam_validation() -> None:
 
         # test that specifying skip doesn't raise error
         client.SensorInfo(f.read(), skip_beam_validation = True)
+
+
+@pytest.mark.parametrize('test_key', ['legacy-2.0', 'single-2.3'])
+def test_original_string(meta: client.SensorInfo, base_name) -> None:
+    meta_path = path.join(PCAPS_DATA_DIR, f"{base_name}.json")
+    with open(meta_path, 'r') as f:
+        assert meta.original_string() == f.read()
+
+
+@pytest.mark.parametrize('metadata_key', ['1_12', '1_12_legacy', '1_13', '1_13_legacy',
+    '1_14_128_legacy', '2_0', '2_0_legacy', '2_1', '2_1_legacy', '2_2', '2_2_legacy', '2_3',
+    '2_3_legacy', '2_4', '2_4_legacy', '2_5', '2_5_legacy', '3_0'])
+def test_updated_string(metadata_base_name) -> None:
+    meta_path = str(Path(METADATA_DATA_DIR) / f"{metadata_base_name}")
+
+    with open(meta_path, 'r') as f:
+        meta = client.SensorInfo(f.read())
+
+    meta.format.columns_per_packet = 40
+    meta.format.fps = 50  # crucial check - fps is a value that is filled in with a default value at parsing time
+    updated_metadata_string = meta.updated_metadata_string()
+
+    assert updated_metadata_string != meta.original_string()
+
+    meta2 = client.SensorInfo(updated_metadata_string, skip_beam_validation=True)
+    assert meta2.format.columns_per_packet == 40
+    assert meta2.format.fps == 50